requirements.txt
requests
mastodon.py
redis
bs4
pythorhead>=0.21.0
python-dotenv
loguru
Pillow